Luffy/requirements.txt
opencv-python
pytesseract
torch
numpy
SpeechRecognition
pyttsx3
ultralytics
google-generativeai
picamera2